Share ResourceFile implementation as a more generic TempFile
