+++ lib/k8s-test-framework/src/resource_file.rs
@@ -1,21 +1,16 @@
-use std::path::{Path, PathBuf};
-use tempfile::{tempdir, TempDir};
+use std::path::Path;
+
+use crate::temp_file::TempFile;
 
 #[derive(Debug)]
-pub struct ResourceFile {
-    dir: TempDir,
-    path: PathBuf,
-}
+pub struct ResourceFile(TempFile);
 
 impl ResourceFile {
     pub fn new(data: &str) -> std::io::Result<Self> {
-        let dir = tempdir()?;
-        let path = dir.path().join("custom.yaml");
-        std::fs::write(&path, data)?;
-        Ok(Self { dir, path })
+        Ok(Self(TempFile::new("custom.yaml", data)?))
     }
 
     pub fn path(&self) -> &Path {
-        self.path.as_path()
+        self.0.path()
     }
 }
